fix(ci): split app lint and test helpers
