+++ packages/app/tests/docker-git/fixtures/event-recorder.ts
@@ -1,0 +1,6 @@
+import { Effect } from "effect"
+
+export const recordEvent = (events: Array<string>, entry: string): Effect.Effect<void> =>
+  Effect.sync(() => {
+    events.push(entry)
+  })
